Refactor: move fetchNightWatchPlanFailedAction to the right place
